+++ src/tiled/mainwindow.cpp
@@ -153,8 +153,7 @@ ExportDetails<Format> chooseExportDetails(const QString &fileName,
 
         QString lastExportedFilePath = session.lastPath(Session::ExportedFile);
 
-        suggestedFilename = lastExportedFilePath
-                            + QLatin1Char('/') + baseName;
+        suggestedFilename = lastExportedFilePath + QLatin1Char('/') + baseName;
         if (!extension.isEmpty())
             suggestedFilename += QLatin1Char('.') + extension;
     }
@@ -2436,8 +2435,8 @@ void MainWindow::exportTilesetAs(TilesetDocument *tilesetDocument)
         fileName = tilesetDocument->tileset()->name();
     }
 
-    SessionOption<QString> lastUsedTilesetExportFilter { "tileset.lastUsedExportFilter" };
-    QString selectedFilter = lastUsedTilesetExportFilter;
+    SessionOption<QString> lastUsedExportFilter { "tileset.lastUsedExportFilter" };
+    QString selectedFilter = lastUsedExportFilter;
     auto exportDetails = chooseExportDetails<TilesetFormat>(fileName,
                                                             tilesetDocument->lastExportFileName(),
                                                             selectedFilter,
@@ -2448,7 +2447,7 @@ void MainWindow::exportTilesetAs(TilesetDocument *tilesetDocument)
     Session &session = Session::current();
 
     session.setLastPath(Session::ExportedFile, QFileInfo(exportDetails.mFileName).path());
-    lastUsedTilesetExportFilter = selectedFilter;
+    lastUsedExportFilter = selectedFilter;
 
     ExportHelper exportHelper;
     SharedTileset exportTileset = exportHelper.prepareExportTileset(tilesetDocument->tileset());