+++ NEWS.md
@@ -26,6 +26,7 @@
 * Fixed tileset tabs to fall back to filename in case of unnamed tilesets (by Sid, #4360)
 * Fixed alpha component of tint color not applying correctly to opaque images (by Roland Helmerichs, #4310)
 * Fixed panning with space bar not always working on first click (with Oval, #4338)
+* Fixed suggesting filename with trailing dot when export filter is unset (by Sid, #4368)
 * Scripting: Added API for custom property types (with dogboydog, #3971)
 * Scripting: Added TileMap.chunkSize and TileMap.compressionLevel properties
 * Scripting: Added tiled.session to read and write session properties (by Kanishka, #4345)